requirements.txt
# Hiro: Multi-Agent Social Governance Environment
# Python dependencies

# Core
pydantic>=2.0.0
numpy>=1.24.0

# API and Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# LLM Client
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
httpx>=0.24.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
